Short circuit the "on" command
The "on" command is special since it doesn't make use of the websocket
connection. It instead sends a wake-on-lan package. This means `ws.connect()`
isn't needed for this and actually just slows down the command because it
causes ws4py to try to establish a connection to a TV that is most likely off,
considering the command being executed.