Drop Python 2.7 support
As far as I can tell, it hasn't worked the last four years following the
changes in commit 0c4ad73. This means the check for which Python version that
is used can be removed since `inspect.getfullargspec` has been available since
Python 3.0.

This also updates the classifiers for the package to the currently maintained
Python versions.