Add setDeviceInfo command
This patch adds a new command, setDeviceInfo, which allows forcing
the TV to use a specific icon (which implies device type) and label
for a specific HDMI input.

Background:

Some LG TVs such as the C4 will drive inputs in a so-called "PC
mode" which means that most of the image processing features become
inactive and unselectable. This happens when the device communicates
itself as a PC or PC-like device (for example, media players like
the Raspberry Pi with LibreELEC will do this). The main issue with
this is that actually disabling PC mode is quite cumbersome and in
some cases totally impossible:

* You have to go to the "Home" dashboard and click the vertical
  ellipsis (three dots) icon.
* Then select "Edit Inputs".
* Then click the icon of the input you want PC mode disabled for.
* Then select one of the non-PC icons.

And you have to repeat this every time the TV is cold booted.

Furthermore, doing this action is simply not possible when HDMI CEC
is enabled for the input (this makes the input not appear at all in
the list of editable inputs).

So, the command is added here which allows the "PC mode" to be
disabled; without the cumbersome menu action and even when HDMI CEC
is enabled. You can use this if you want to attach a PC or PC-like
device but still want to take advantage of image processing.

Usage:

    lgtv --ssl setDeviceInfo <id> <icon> <label>

Usage example:

    lgtv --ssl setDeviceInfo HDMI_4 HDMI_4.png "My Kodi"

Input IDs and options for the "icon" argument can be listed with:

    lgtv --ssl listInputs

The "icon" argument only needs the base filename, not the full URL.
The "label" argument has to be enclosed in quotes if it contains
any spaces.